RISC-V: interrupts: fix Zifencei emulation on rv64
Zifencei emulaton was added for bare bones 32-bit targets, but the code
is there on 64-bit too. The instruction skip at the end of the function
skips 4 bytes, the size of an instruction on both rv32 and rv64. The
instruction decoder however assumes instructions are size(long) bytes,
which is 8 > 4 on rv64. Fix this.

Link: https://lore.barebox.org/20211203063340.1026512-1-ahmad@a3f.at

+++ arch/riscv/cpu/interrupts.c
@@ -122,7 +122,7 @@ unsigned long handle_trap(struct pt_regs *regs)
 		goto skip;
 
 	if (regs->cause == 2) { /* illegal instruction */
-		switch(*(unsigned long *)regs->epc) {
+		switch(*(u32 *)regs->epc) {
 		case 0x0000100f: /* fence.i */
 			goto skip;
 		default:
